+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Stacked bars layer and example
 - Bubble chart layer and example, example includes functionality to map model values to colors from gradient bar
 - Scatter chart layer and example
+- Bars layer, +/- bars with dynamic gradient example 
 
 ### Removed
 - Bars example with horizontal axis - now included in variable axes example